feat(architecture): ✨ 增强指标聚合能力、新增 Timeout 监控与 CLI 诊断插件
- Rust Agent: 新增 MetricsAggregator 环形缓冲区，支持 avg/max/min/p95 统计与趋势检测
- SDK: 新增 TimeoutSubject、支持 per-metric 开关与采集间隔配置、重构 proxy_thread 路由
- CLI: 新增 monitor-memory 和 timers 插件、MemoryGraph TUI 组件、InspectorError 错误分级
- CI: 添加 GitHub Actions 工作流（lint + TS 测试 + Rust 测试）
- 补充全面单测覆盖（e2e、plugins、services、collector、subjects）

+++ agent/src/ipc/http/endpoints/heartbeat.rs
@@ -21,7 +21,7 @@ pub struct HeartbeatRouter {
 
 #[derive(Deserialize)]
 pub struct HeartbeatRequest {
-    process_id: u16,
+    process_id: u32,
 }
 
 impl BaseRouter for HeartbeatRouter {
